refactored msw handlers to mocks directory to avoid handler duplication code
