+++ client/src/tests/ProductPage.test.js
@@ -5,78 +5,12 @@ import { MemoryRouter, Route, Routes} from 'react-router-dom';
 import { CartProvider} from "../features/Cart/CartContext";
 import { CartProviderPopUp } from "../features/Cart/CartPopUpContext";
 import { SelectedProductProvider } from "../features/Products/SelectedProductContext";
-import {setupServer} from 'msw/node';
-import {rest} from 'msw';
+
 import { ProductsProvider } from "../features/Products/ProductsContext";
 import userEvent from '@testing-library/user-event';
 import Login from "../features/Login/Login";
 //first intercept on mount call with msw to se selected product
-const server = setupServer(
-    // Define mock responses for your API endpoints
-    rest.get('http://localhost:4000/products/1', (req, res, ctx) => {
-      return res(
-        ctx.status(200),
-        ctx.json([
-          {
-            id: 1,
-            product_name: 'Elegant Gold Bangle',
-            image_url: 'http://localhost:4001/elegant-gold-bangle',
-            product_description: 'Elevate your style with this exquisite gold bangle, weighing a delicate 10 grams. Its intricate design and comfortable fit make it a perfect accessory for any occasion.',
-            price: 1100
-          }
-        ])
-      );
-    }),
-  
-    rest.get('http://localhost:4000/products/categories', (req, res, ctx) => {
-      return res(
-        ctx.status(200),
-        ctx.json([
-          {
-            categories: 'bangles'
-          }
-        ])
-      );
-    }),
-  
-    rest.get('http://localhost:4000/carts/1', (req, res, ctx) => {
-      return res(
-        ctx.status(200),
-        ctx.json([
-          {
-            id: 1,
-            user_id: 1,
-            product_id: 1,
-            quantity: 2,
-          }
-        ])
-      );
-    }),
-  
-    rest.get('http://localhost:4000/carts/cart-total', (req, res, ctx) => {
-      // Access query parameters using req.url.searchParams
-      const userId = req.url.searchParams.get('user_id');
-      
-      // Return the response based on the userId
-      if (userId === '1') {
-        return res(
-          ctx.status(200),
-          ctx.json([
-            {
-              total_cost: 1100
-            }
-          ])
-        );
-      } 
-    })
-  );
-  
-  // Start the MSW server
-  beforeAll(() => server.listen());
-  
-  // Stop the MSW server after the tests
-  afterAll(() => server.close());
-  
+
 
 test('renders selected product information correctly', async() =>{
     render(
@@ -148,7 +82,7 @@ test('cart opened when clicked' , async () => {
             const cartComponent = screen.getByRole('heading', {name: 'Cart'});
             expect(cartComponent).toBeInTheDocument()
          })
-         screen.debug()
+      
 })
 
 //test route to login page
@@ -181,7 +115,6 @@ test('test route to login page', async () => {
             user.click(logInText);
             const logInHeading = screen.getByRole('heading', {name: 'Login'})
             expect(logInHeading).toBeInTheDocument();
-            screen.debug()
-
+            
         })
     })
